moved fetchers to new, common module
