+++ packages/extension/src/popup.tsx
@@ -1,13 +1,21 @@
 import React from "react";
-import { Paragraph } from "@paperpod/ui"
-import { useUrl } from "./effects"; 
+import { Button, Paragraph } from "@paperpod/ui";
+import { useUrl } from "./effects";
+
+const postArticle;
 
 export const Popup = () => {
+  const url = useUrl();
 
-  const url = useUrl()
-  return <>
-    <h1>Paperpod</h1>
-  <Paragraph>
-    current url {url}
-    </Paragraph></>
+  const onClick = () => {
+    window.close();
+  };
+
+  return (
+    <>
+      <h1>Paperpod</h1>
+      <Paragraph>current url {url}</Paragraph>
+      <Button onClick={onClick}>Close me</Button>
+    </>
+  );
 };
