+++ packages/frontend/src/fetchers/bearer.ts
@@ -1,0 +1,6 @@
+export const bearer = (token: string) => ({
+  headers: {
+    authorization: "Bearer " + token,
+    "Content-Type": "application/json",
+  },
+});
