+++ packages/frontend/src/fetchers/fetchers.ts
@@ -1,0 +1,8 @@
+import * as article from "./articleFetchers";
+import * as auth from "./authFetchers";
+import * as player from "./playerFetchers";
+export const fetchers = {
+  article,
+  auth,
+  player,
+};
