+++ packages/frontend/src/fetchers/playerFetchers.ts
@@ -1,4 +1,4 @@
-import { bearer } from "../../helpers/bearer";
+import { bearer } from "./bearer";
 
 export const fetchFile = (article_id: string, token: string) =>
   fetch(`/api/files/${article_id}`, bearer(token));
