+++ packages/web/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@paperpod/common": "0.0.1",
     "@paperpod/ui": "0.0.1",
+    "@paperpod/frontend": "0.0.1",
     "@testing-library/jest-dom": "^5.12.0",
     "next": "^9.4.0",
     "next-transpile-modules": "^6.0.0",
